requirements.txt
streamlit
pandas
google-cloud-bigquery
google-cloud-bigquery-storage
google-auth
google-auth-oauthlib
google-auth-httplib2
db-dtypes